Fix an oversight with validations and switch to jnunemaker's Validatable.
When planning the initial API, I intended on using it much like a Ruby argument
list. I forgot that the API for the validations is basically
`validation :field, ARGV`. So, I changed the API for it in our platform. Also,
the official Validatable isn't very feature-complete, so we'll be using a
different branch of it (jnunemaker's branch).

As a result of the above changes, I updated the README.md file.