Enforce uniqueness of DataObjectDefinitions
To enforce uniqueness of DataObjectDefinitions in the database, I'm setting the unique ID to the object type. I also considered looking at the database on save and leaving the default ID, but this should work better with
replication and such. Or at least, that's what a crapton of blog posts said. If it ends up being a problem, we can also check the database before saving. Though I think it'll work.