Skeleton for DataObject.
Trees in CouchDB are bizzare to say the least. It makes sense, and really is a lot more awesome.
For starters, you can totally move an object just by changing the document to be moved. And we
wouldn't have to traverse the tree since we'd have a copy of it. But there is the storage
overhead with that... So I dunno. I think storing the path would be the best use of CouchDB's
features, though we can still store the relationship in a key like _parent_id. Though that
means one HTTP call per level of the tree. Oh, and doing it this way would allow us to filter
all of the items under a specific portion of the tree... So no more categories hack for pulling
all of the events in a city.

Seriously though, CouchDB's wiki needs help. It's really not documented very well how to use
include_docs or several other features. Or at least not in a way that makes sense to me.

Ahem, anyways. Apologies for the insanely long commit message. May it help you explain CouchDB
to yourself like it did me ;)