Add a reduce function to the path. Returns number of siblings.
If two documents have a path of [], they would reduce into [], 2.
Put another way, if documents have a path of ["country/US",
"region/TX", "locality/Austin"], we can figure out how many objects
are inside Austin. We can further refine the results in subclasses
so we can get events, places, etc. An issue would be that if we
wanted to use startkey/endkey to say, get all events for a place,
we'd need to sum the result of the reduce function lest there be
children nodes.